refactor: extract shared AmountWithLabel and EquationRow
