+++ src/components/TaxDetails/TaxSummaryCard/TaxSummaryCardDesktop.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next'
 
-import type { TaxSummary, TaxSummarySection } from '@schemas/taxEstimates/summary'
+import type { TaxSummary } from '@schemas/taxEstimates/summary'
 import { tConditional } from '@utils/i18n/conditional'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useFullYearProjection } from '@providers/TaxEstimatesRouteStore/TaxEstimatesRouteStoreProvider'
@@ -9,47 +9,10 @@ import Plus from '@icons/Plus'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { MoneySpan } from '@ui/Typography/MoneySpan'
 import { Span } from '@ui/Typography/Text'
-import { Badge, BadgeSize, BadgeVariant } from '@components/Badge/Badge'
+import { EquationRow } from '@components/TaxDetails/TaxSummaryCard/TaxSummaryCardEquation'
 
 type TaxSummaryCardDesktopProps = {
   data: TaxSummary
-}
-
-type AmountWithLabelProps = {
-  amount: number
-  label: string
-  emphasis?: boolean
-}
-
-const AmountWithLabel = ({ amount, label, emphasis }: AmountWithLabelProps) => (
-  <VStack className='Layer__TaxSummaryCard__AmountWithLabel' gap='2xs' align='start'>
-    <MoneySpan size='lg' weight={emphasis ? 'bold' : undefined} amount={amount} />
-    <Badge size={BadgeSize.SMALL} variant={BadgeVariant.NEUTRAL}>{label}</Badge>
-  </VStack>
-)
-
-type SectionEquationProps = {
-  section: TaxSummarySection
-}
-
-const SectionEquation = ({ section }: SectionEquationProps) => {
-  const { t } = useTranslation()
-  return (
-    <HStack className='Layer__TaxSummaryCard__Equation' gap='md'>
-      <AmountWithLabel
-        amount={section.taxesOwed}
-        label={t('taxEstimates:label.taxes_owed', 'Taxes Owed')}
-        emphasis
-      />
-      <Span className='Layer__TaxSummaryCard__Operator' size='lg' variant='subtle'>=</Span>
-      <AmountWithLabel amount={section.total} label={t('common:label.total', 'Total')} />
-      <Span className='Layer__TaxSummaryCard__Operator' size='lg' variant='subtle'>-</Span>
-      <AmountWithLabel
-        amount={section.taxesPaid}
-        label={t('taxEstimates:label.taxes_paid', 'Taxes paid')}
-      />
-    </HStack>
-  )
 }
 
 const PlusCircle = () => (
@@ -109,7 +72,7 @@ export const TaxSummaryCardDesktop = ({ data }: TaxSummaryCardDesktopProps) => {
           key={`${section.label}-body`}
           className='Layer__TaxSummaryCard__Cell Layer__TaxSummaryCard__Cell--bordered'
         >
-          <SectionEquation section={section} />
+          <EquationRow section={section} size='lg' />
         </VStack>
       ))}
     </div>
